+++ pystruct3d/preprocessing/__init__.py
@@ -1,3 +1,6 @@
 from pystruct3d.preprocessing.alignment import align_to_axes as align_to_axes
+from pystruct3d.preprocessing.crop import crop_roi as crop_roi
 from pystruct3d.preprocessing.labels import filter_ids as filter_ids
 from pystruct3d.preprocessing.labels import labels_to_color as labels_to_color
+from pystruct3d.preprocessing.pca import rotate_by_pca as rotate_by_pca
+from pystruct3d.preprocessing.pca import simple_pca as simple_pca